pyproject.toml
[project]
name = "open-wearables"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=45.0.4",
    "passlib>=1.7.4",
    "pydantic-settings>=2.10.1",
    "email-validator>=2.2.0",
    "psycopg>=3.2.9",
    "sqlalchemy>=2.0.43",
    "fastapi>=0.116.1",
    "fastapi-cli>=0.0.8",
    "sqladmin[full]>=0.21.0",
    "celery>=5.5.3",
    "flower>=2.0.1",
    "redis>=7.0.1",
    "sentry-sdk[fastapi]>=2.42.1",
    "python-multipart>=0.0.20",
    "python-jose[cryptography]>=3.5.0",
    "httpx>=0.28.1",
]

[dependency-groups]
code-quality = [
    "pre-commit>=4.3.0",
    "ruff>=0.12.10",
    "ty>=0.0.1a19",
]
dev = [
    "alembic>=1.16.4",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.2.1",
]

[tool.ruff]
line-length = 120
target-version = 'py313'

[tool.ruff.lint]
select = [
    "I",     # isort
    "F",     # pyflakes
    "FAST",  # FastApi
    "ANN",   # flake8-annotations
    "ASYNC", # flake8-async
    "COM",   # flake8-commas
    "T10",   # flake8-debugger
    "PT",    # flake8-pytest-style
    "RET",   # flake8-return
    "SIM",   # flake8-simplify
    "N",     # pep8-naming
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
]
ignore = [
    "ANN002", # missing-type-args
    "ANN003", # missing-type-kwargs
    "ANN204", # missing-return-type-special-method
    "ANN401", # any-type
    "RET503", # implicit-return
]


[build-system]
requires = ["uv_build"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-root = ""
module-name = "app"
